Added model fuzzer, and test that uses it.
  Works, but very slow: a 100 item run takes 50-65s. Printing timing info revealed that
only about 1/3rd of the time is spent running the test validation. The rest is presumably
the fuzzer. The shrinker never seems to come into play (which makes sense since all
tests pass). This seems to be a known issue with fuzzers:

  https://github.com/elm-community/elm-test/issues/237

  With that said, I'll put this branch on hold. The code is as-is, as in not in a state
of which I'm particularly proud of but whatever :)